Backend/requirements.txt
bcrypt==4.3.0
fastapi==0.115.12
python-jose==3.4.0
uvicorn==0.34.0
SQLAlchemy==2.0.40
python-dotenv==1.0.1
passlib==1.7.4
